TSCH and CSMA: consistent seqno handling

+++ core/net/mac/csma/csma-output.c
@@ -483,7 +483,7 @@ csma_output_packet(mac_callback_t sent, void *ptr)
   struct packet_queue *q;
   struct neighbor_queue *n;
   static uint8_t initialized = 0;
-  static uint16_t seqno;
+  static uint8_t seqno;
   const linkaddr_t *addr = packetbuf_addr(PACKETBUF_ADDR_RECEIVER);
 
   if(!initialized) {
